pyproject.toml
[project]
name = "enrichment-engine"
version = "0.1.0"
description = ""
authors = [
    {name = "Sandipan"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1,<0.117.0",
    "uvicorn>=0.35.0,<0.36.0", 
    "pandas>=2.3.1,<3.0.0",
    "requests>=2.32.5,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",  # Updated from 'dotenv'
    "google-generativeai>=0.8.5,<0.9.0",
    "google-genai>=1.31.0,<2.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "sqlalchemy>=2.0.43,<3.0.0",
    "geopy>=2.4.1,<3.0.0",
    # Cloud-specific dependencies
    "google-cloud-secret-manager>=2.16.0,<3.0.0",
    "google-cloud-logging>=3.8.0,<4.0.0",
    "google-auth>=2.17.0,<3.0.0",
    "google-cloud-core>=2.3.0,<3.0.0",
    "pg8000 (>=1.31.4,<2.0.0)",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.0.0",
    "black>=23.7.0,<24.0.0",
    "flake8>=6.0.0,<7.0.0",
    "mypy>=1.5.0,<2.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"

